Trends on weight from Database

+++ ModelLibrary/ControlSchedules.cs
@@ -9,7 +9,7 @@ namespace ModelLibrary
             
         }
 
-        public ControlSchedules(int control_ID, DateTime time, DateTime weightControlTime, double weight, string receiptForTest, string kegTest, double ludKoncentration, double mipMA, string signature, string note)
+        public ControlSchedules(int control_ID, DateTime time, DateTime weightControlTime, int weight, string receiptForTest, string kegTest, double ludKoncentration, double mipMA, string signature, string note)
         {
             Control_ID = control_ID;
             Time = time;
@@ -26,7 +26,7 @@ namespace ModelLibrary
         public int Control_ID { get; set; } //PK
         public DateTime Time { get; set; }
         public DateTime WeightControlTime { get; set; } //Tidspunktet hvor vægten blev kontroleret
-        public double Weight { get; set; }
+        public int Weight { get; set; }
         public string ReceiptForTest { get; set; }
         public string KegTest { get; set; }
         public double LudKoncentration { get; set; }
